Moved DjangoWhiteNoise into handlers.
